% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{numeric_covariates}
\alias{numeric_covariates}
\title{Turn categorical covariates into numeric values}
\usage{
numeric_covariates(directory, filename, sheetindex = 1, types, covariate_name)
}
\arguments{
\item{directory}{The directory of the excel file. This must be the full directory name. An easy way to find this is to set the directory using RStudio's "Files" tab and copying and pasting the directory name from the console. Enter as a string (with quotations).}

\item{filename}{The excel file's name. Make sure to include the file extension, in this case .xlsx. Enter as a string.}

\item{sheetindex}{The sheet to be read into R. Default = 1.}

\item{types}{A vector of strings indicating the types of each of the columns in the excel sheet being read into R. The data types must be "numeric" or "character".}

\item{covariate_name}{The name of the covariate that takes the numeric values. Must match the header of the column with the covariates. Enter as a string.}
}
\description{
This function takes the categorical covariate covariate_name within an excel file and converts each level of the variable to a numeric value. Based on the as.numeric function, the categorical variables are assigned numbers in alphabetical order starting with 1. The new covariate values will be added as a column to the original excel sheet and a new file will be created in the directory specified.
}
\examples{
types <- c( "numeric", "numeric", "character", "numeric", "numeric", "numeric", "numeric", "character", "numeric", "numeric", "numeric" )
numeric_covariates ( directory = "~/Desktop/Clinical Pharmacology/Kyelee Fitts/Docetaxel_phoenix", filename = "popPK data.xlsx", sheetindex = 1, types = types, covariate_name = "Race" )
numeric_covariates()
}
\keyword{categorical}
\keyword{covariates}
\keyword{numeric}
