% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{times_as_zeros}
\alias{times_as_zeros}
\title{Make negative times zeros.}
\usage{
times_as_zeros(directory, filename, sheetindex, types, time_column)
}
\arguments{
\item{directory}{The directory of the excel file. This must be the full directory name. An easy way to find this is to set the directory using RStudio's "Files" tab and copying and pasting the directory name from the console. Enter as a string (with quotations).}

\item{filename}{The excel file's name. Make sure to include the file extension, in this case .xlsx. Enter as a string.}

\item{sheetindex}{The sheet to be read into R. Default = 1.}

\item{types}{A vector of strings indicating the types of each of the columns in the excel sheet being read into R. The data types must be "numeric" or "character".}

\item{time_column}{The heading of the column of the column containing the times. Must match the header of the column of the Excel file. Enter as string.}
}
\description{
This function takes all times in the column time_column within a sheet of an excel file and converts the negative time values to zeros.
}
\examples{
types <- c( "numeric", "numeric", "character", "numeric", "numeric", "numeric", "numeric", "character", "numeric", "numeric", "numeric" )
times_as_zeros ( "C:/Users/FiggLab4357/Desktop/Clinical Pharmacology/Kyelee Fitts/Docetaxel_phoenix", "PopPK Data.xlsx", sheetindex = 1, types, "DTX_Time" )
times_as_zeros()
}
\keyword{negative}
\keyword{times}
\keyword{zeros}
